Finished basic Codeberg implementation and Gitea base API split
