--- src/main/kotlin/me/modmuss50/mpp/platforms/gitea/GiteaCompatible.kt
@@ -1,242 +1,0 @@
-package me.modmuss50.mpp.platforms.gitea
-
-import me.modmuss50.mpp.GiteaCompatiblePublishResult
-import me.modmuss50.mpp.Platform
-import me.modmuss50.mpp.PlatformOptions
-import me.modmuss50.mpp.PlatformOptionsInternal
-import me.modmuss50.mpp.PublishContext
-import me.modmuss50.mpp.PublishModTask
-import me.modmuss50.mpp.PublishOptions
-import me.modmuss50.mpp.PublishResult
-import me.modmuss50.mpp.PublishWorkAction
-import me.modmuss50.mpp.PublishWorkParameters
-import me.modmuss50.mpp.ReleaseType
-import org.gradle.api.Task
-import org.gradle.api.file.RegularFileProperty
-import org.gradle.api.logging.Logger
-import org.gradle.api.provider.Property
-import org.gradle.api.provider.Provider
-import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.InputFile
-import org.gradle.api.tasks.Optional
-import org.gradle.api.tasks.TaskProvider
-import org.jetbrains.annotations.ApiStatus.Internal
-import java.net.URI
-import java.util.Locale
-import javax.inject.Inject
-import kotlin.random.Random
-
-interface GiteaCompatibleOptions : PlatformOptions, PlatformOptionsInternal<GiteaCompatibleOptions> {
-    @get:InputFile
-    @get:Optional
-    override val file: RegularFileProperty
-
-    /**
-     * "owner/repo"
-     */
-    @get:Input
-    val repository: Property<String>
-
-    /**
-     * Specifies the commitish value that determines where the Git tag is created from. Can be any branch or commit SHA.
-     */
-    @get:Input
-    @get:Optional
-    val commitish: Property<String>
-
-    @get:Input
-    val tagName: Property<String>
-
-    @get:Input
-    val apiEndpoint: Property<String>
-
-    /**
-     * Specifies the display name for the custom host. For example, Codeberg.
-     */
-    @get:Input
-    @get:Optional
-    val hostDisplayName: Property<String>
-
-    /**
-     * Specifies a custom brand color for Discord embeds. Useful for specific hosts.
-     */
-    @get:Input
-    @get:Optional
-    val hostBrandColor: Property<Int>
-
-    @get:Input
-    val allowEmptyFiles: Property<Boolean>
-
-    @get:InputFile
-    @get:Optional
-    @get:Internal
-    val releaseResult: RegularFileProperty
-
-    @get:Input
-    @get:Internal
-    val hostType: Property<GiteaCompatiblePlatform>
-
-    override fun setInternalDefaults() {
-        tagName.convention(version)
-        allowEmptyFiles.convention(false)
-    }
-
-    fun host(uri: URI) {
-        apiEndpoint.convention("$uri/api/v1")
-    }
-
-    fun from(other: GiteaCompatibleOptions) {
-        super.from(other)
-        repository.convention(other.repository)
-        commitish.convention(other.commitish)
-        tagName.convention(other.tagName)
-        apiEndpoint.convention(other.apiEndpoint)
-        allowEmptyFiles.convention(other.allowEmptyFiles)
-        releaseResult.convention(other.releaseResult)
-        hostType.convention(other.hostType)
-    }
-
-    fun from(other: Provider<GiteaCompatibleOptions>) {
-        from(other.get())
-    }
-
-    fun from(other: Provider<GiteaCompatibleOptions>, publishOptions: Provider<PublishOptions>) {
-        from(other)
-        from(publishOptions.get())
-    }
-
-    /**
-     * Publish to an existing release, created by another task.
-     */
-    fun parent(task: TaskProvider<Task>) {
-        val publishTask = task.map { it as PublishModTask }
-        releaseResult.set(publishTask.flatMap { it.result })
-
-        val options = publishTask.map { it.platform as GiteaCompatibleOptions }
-        if (options.get().hostType.get() != hostType.get()) { // May not be necessary, but should reduce confusion.
-            throw IllegalStateException("Unable to make a ${options.get().hostType.get().friendlyString} instance a parent of a ${hostType.get().friendlyString} instance")
-        }
-
-        version.set(options.flatMap { it.version })
-        version.finalizeValue()
-        changelog.set(options.flatMap { it.changelog })
-        changelog.finalizeValue()
-        type.set(options.flatMap { it.type })
-        type.finalizeValue()
-        displayName.set(options.flatMap { it.displayName })
-        displayName.finalizeValue()
-        repository.set(options.flatMap { it.repository })
-        repository.finalizeValue()
-        commitish.set(options.flatMap { it.commitish })
-        commitish.finalizeValue()
-        tagName.set(options.flatMap { it.tagName })
-        tagName.finalizeValue()
-
-        // Include this here to make sure that different hosts may resolve correctly when parenting. This is not included in other platforms.
-        apiEndpoint.set(options.flatMap { it.apiEndpoint })
-        apiEndpoint.finalizeValue()
-    }
-}
-
-abstract class GiteaCompatible @Inject constructor(name: String) : Platform(name), GiteaCompatibleOptions {
-    override fun publish(context: PublishContext) {
-        val files = additionalFiles.files.toMutableList()
-
-        if (file.isPresent) {
-            files.add(file.get().asFile)
-        }
-
-        if (files.isEmpty() && !allowEmptyFiles.get()) {
-            throw IllegalStateException("No files to upload to ${capitalizedName()}.")
-        }
-
-        context.submit(UploadWorkAction::class) {
-            it.from(this)
-        }
-    }
-
-    override fun dryRunPublishResult(): PublishResult {
-        val hostDisplayName = hostDisplayName.getOrElse(hostType.get().friendlyString)
-        val brandColor = hostBrandColor.getOrElse(hostType.get().defaultBrandColor)
-
-        return GiteaCompatiblePublishResult(
-            repository = repository.get(),
-            releaseId = 0,
-            url = "https://github.com/modmuss50/mod-publish-plugin/dry-run?random=${Random.nextInt(0, 1000000)}",
-            title = announcementTitle.getOrElse("Download from $hostDisplayName"),
-            brandColor = brandColor,
-        )
-    }
-
-    override fun printDryRunInfo(logger: Logger) {
-    }
-
-    fun capitalizedName(): String {
-        return name.replaceFirstChar { if (it.isLowerCase()) it.titlecase(Locale.ROOT) else it.toString() }
-    }
-
-    interface UploadParams : PublishWorkParameters, GiteaCompatibleOptions
-
-    abstract class UploadWorkAction : PublishWorkAction<UploadParams> {
-        override fun publish(): PublishResult {
-            with(parameters) {
-                val hostDisplayName = hostDisplayName.getOrElse(hostType.get().friendlyString)
-                val brandColor = hostBrandColor.getOrElse(hostType.get().defaultBrandColor)
-
-                val api = GiteaCompatibleApi(accessToken.get(), apiEndpoint.get(), repository.get())
-                val (release, created) = getOrCreateRelease(api)
-
-                val files = additionalFiles.files.toMutableList()
-
-                if (file.isPresent) {
-                    files.add(file.get().asFile)
-                }
-
-                val noneUnique = files.groupingBy { it.name }.eachCount().filter { it.value > 1 }
-                if (noneUnique.isNotEmpty()) {
-                    val noneUniqueNames = noneUnique.keys.joinToString(", ")
-                    throw IllegalStateException("${hostType.get().friendlyString} file names must be unique within a release, found duplicates: $noneUniqueNames")
-                }
-
-                for (file in files) {
-                    api.uploadAsset(release, file)
-                }
-
-                if (created) {
-                    // Publish the release after all assets are uploaded.
-                    api.publishRelease(release)
-                }
-
-                return GiteaCompatiblePublishResult(
-                    repository = repository.get(),
-                    releaseId = release.id,
-                    url = release.htmlUrl,
-                    title = announcementTitle.getOrElse("Download from $hostDisplayName"),
-                    brandColor = brandColor,
-                )
-            }
-        }
-
-        data class ReleaseResult(val release: GiteaCompatibleApi.Release, val created: Boolean)
-
-        private fun getOrCreateRelease(api: GiteaCompatibleApi): ReleaseResult {
-            with(parameters) {
-                if (releaseResult.isPresent) {
-                    val result = PublishResult.fromJson(releaseResult.get().asFile.readText()) as GiteaCompatiblePublishResult
-                    return ReleaseResult(api.getRelease(result.releaseId), false)
-                }
-
-                val metadata = GiteaCompatibleApi.CreateRelease(
-                    body = changelog.orNull,
-                    draft = true,
-                    name = displayName.get(),
-                    prerelease = type.get() != ReleaseType.STABLE,
-                    tagName = tagName.get(),
-                    targetCommitish = commitish.get(),
-                )
-
-                return ReleaseResult(api.createRelease(metadata), true)
-            }
-        }
-    }
-}
