+++ src/main/kotlin/me/modmuss50/mpp/platforms/gitea/base/GiteaCompatibleApi.kt
@@ -1,11 +1,10 @@
-package me.modmuss50.mpp.platforms.gitea
+package me.modmuss50.mpp.platforms.gitea.base
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import me.modmuss50.mpp.networking.DefaultHttpImpl
-import me.modmuss50.mpp.networking.DefaultHttpImpl.jsonErrorFactory
 import me.modmuss50.mpp.networking.HttpConfig
 import me.modmuss50.mpp.networking.HttpContext
 import me.modmuss50.mpp.networking.MultipartBodyBuilder
@@ -19,7 +18,7 @@ class GiteaCompatibleApi(
 ) {
     companion object {
         val giteaExceptionFactory =
-            jsonErrorFactory<ErrorResponse> {
+            DefaultHttpImpl.jsonErrorFactory<ErrorResponse> {
                 it.message
             }
